pyproject.toml
[project]
name = "dc2host"
version = "0.0.1"
dependencies = ["docker", "flask", "GitPython"]
authors = [
  { name="Vlad Folts", email="vladfolts@gmail.com" },
]
description = "VS Code Dev Containers integration with a host system"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"